django/requirements.txt
Django==4.2.16
djangorestframework
django-cors-headers
mysqlclient